Add a waste matric and use that to find the best solution
Calculate a waste for each solution. Waste is calculated as, for each input, the
cost of spending an input now - cost of spending an input later, + the
excess. We then search for all solutions and find the one that has the least waste.

Additional optimizations have been made with the waste calculation:
Any branch that is more wasteful than our best solution is ignored
and the branch is cut