Add sweepwallet RPC for wallet
_Motivation_
Currently, the wallet uses a fSubtractFeeAmount (SFFA) flag on the
recipients objects for all forms of sending calls. According to the
commit discussion, this flag was chiefly introduced to permit sweeping
without manually calculating the fees of transactions. However, the flag
leads to unintuitive behavior and makes it more complicated to test
many wallet RPCs exhaustively. We proposed to introduce a dedicated
`sweepwallet` RPC with the intention to cover this functionality.

Since the proposal, it was discovered in further discussion that our
proposed `sweepwallet` rpc and SFFA have subtly different scopes of
operation.
• sweepwallet:
  Use _given UTXOs_ to pay a destination the remainder after fees.
• SFFA:
  Use a _given budget_ to pay an address the remainder after fees.

While `sweepwallet` will simplify cases of spending a given set of
UTXOs such as paying the value from one or more specific UTXOs, emptying
a wallet, or burning dust, we realized that there are some cases in
which `sweepwallet` is used to pay other parties from a limited budget,
which can often lead to the creation of change outputs. This cannot be
easily replicated using `sweepwallet` as it would require manual
computation of the appropriate change amount.

As such, sweepwallet cannot replace all uses of sffa, but it still has a
different use case and will aid in simplifying some wallet calls and
numerous wallet tests.

_Sweepwallet call details_
The proposed sweepwallet call builds a transaction from a specific
subset of the wallet's UTXO pool (by default all of them) and assigns
the funds to one or more receivers. Receivers can either be specified
with a given amount or receive an equal share of the remaining
unassigned funds. At least one recipient must be provided without
assigned amount to collect the remainder. The `sweepwallet` call will
never create change. The call has a `sendmax` option that changes the
default behavior of spending all UTXOs ("no UTXO left behind"), to
maximizing the output amount of the transaction by skipping uneconomic
UTXOs. The `sendmax` option is incompatible with providing a specific
set of inputs.