Update rpc_fundrawtransaction.py
`test_option_feerate`: Reduce explicit fees to stay within the new maxtxfee default. This test is verifying that setting an explicit fee rate works as expected, but the actual fee amount/rate being specified is irrelevant to this test.

`test_22670`: Have to manually override the new default maxtxfee in order for the test to proceed with its original amounts. Same for
maxfeerate in the testmempoolaccept call. Check with @achow101 that this change doesn't interfere with what this test is trying to verify?